Move contents of author-errors.ts to authors-errors.ts and similar for identifier
