+++ src/authors-errors.ts
@@ -1,6 +1,19 @@
 import { AuthorsType } from 'src/types'
 import { messageErrorType, getMyErrorsArray } from 'src/store/validator'
-import { authorHasErrors } from 'src/author-errors'
+
+export const authorHasErrors = (index:number): boolean => {
+    const errors = getMyErrorsArray([
+        { instancePath: `/authors/${index}/given-names` },
+        { instancePath: `/authors/${index}/name-particle` },
+        { instancePath: `/authors/${index}/family-names` },
+        { instancePath: `/authors/${index}/name-suffix` },
+        { instancePath: `/authors/${index}/email` },
+        { instancePath: `/authors/${index}/affiliation` },
+        { instancePath: `/authors/${index}/orcid` }
+    ])
+
+    return errors.length > 0
+}
 
 export const authorsErrors = (authors:AuthorsType) => {
     const allMessages = getMyErrorsArray([
